Changes due to change of spec from Jim.
Dear All - there is a specification error in the Loan Entity.

The current spec says that the loan constructor takes an id parameter which should be a positive interger, and throws an IllegalArgumentException if that parameter is less than or equal to zero. This is incorrect.

In my new implementation, the constructor does not take an id parameter, and the Loan constructor returns a new loan with a default id of zero. This is the correct behaviour.

The current spec says that the loan 'commit' method takes no parameters. However, the interface file shows that the commit method takes a loan id parameter, and this is the specification used in the reference implementation. The id parameter should be a positive integer.

Here are the revised specifications for the constructor and the commit methods in the Loan entity:

Constructor: Loan

Parameters: book:IBook, borrower: IMember, borrowDate:Date, dueDate:Date

throws IllegalArgumentException if:

any of book, borrower, borrowDate, or dueDate are null

dueDate is less than borrowDate

commit: void

Parameters: loanId:integer

sets the current state of the Loan to CURRENT

calls book.borrow with itself as parameter

calls borrower.addloan with itself as parameter

throws IllegalArgumentException if:

loanId is less than or equal to zero

throws a RuntimeException if:

the loan’s current LoanState is not PENDING

I apologise for my oversight in not revising the specification to match the current implementation - and warn that there may be other such errors (although I hope not). In general, where there is a conflict between the specification and the supplied interface files, take the supplied interface files as definitive - these are the interfaces actually used in the reference implementation.Please let me know if you come across any other anomalies or errors.

best regards

Jim